requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
sse-starlette
aioredis